refactor: extract shared route helper and use microtask confirmation for sticky directory
